python/requirements.txt
avro-python3==1.10.2
Faker==18.11.1
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
